schema: Add docker overrides for entrypoint and user
We have two issues in CI right now that both require overriding
default docker-run behavior.

1) A container with a non-root user may not be able to save artifacts
If the worker is run as root, then /archive is owned by root and the
the container user won't be able to write to it.

2) A container may have an entrypoint we don't want to run
A container like:

 https://github.com/zephyrproject-rtos/docker-image

Includes an entrypoint we don't really want to run in our CI.

Signed-off-by: Andy Doan <andy@foundries.io>