runner: Try to do better for gilab
gitlab secrets might also be for the public gitlab server. e.g. you
have an LmP manifest that refers to a private gilab.com repository.

There's currently not a fool-proof way to know what gitlabtok is for,
but this tries to do the right thing and will, in general, do the right
thing.

Signed-off-by: Andy Doan <andy@foundries.io>