api: Fix run tailing with no new data
Older versions of gunicorn allowed sendfile to read from a file
descriptor that had no data to read. It happened to default to
reading the entire file.

The version we just updated to no longer allows this and throws:

    ValueError: count must be a positive integer (got 0)

        at _check_sendfile_params (/usr/lib/python3.9/socket.py:461)
        at _sendfile_use_sendfile (/usr/lib/python3.9/socket.py:347)
        at sendfile (/usr/lib/python3.9/socket.py:483)
        at sendfile (/usr/lib/python3.9/site-packages/gunicorn/http/wsgi.py:375)
        at write_file (/usr/lib/python3.9/site-packages/gunicorn/http/wsgi.py:385)
        at handle_request (/usr/lib/python3.9/site-packages/gunicorn/workers/sync.py:182)

The "fix" didn't quite work because our tailing code seems to fall
over when a 0byte response is sent. This fixes things for now and
then we can go back an improve this later.

Signed-off-by: Andy Doan <andy@foundries.io>