Add plumbing for worker JWT auth
In order to make workers more flexible, we are introducing the notion
of JWT authentication of workers. Workers with a trusted JWT can
be auto-registered and enlisted into JobServ.

The server can be configured with x509 certificates that it may trust.
Then operators can create workers with JWT's signed by the private
key.

The only special attribute in the x509 cert is the handling of the `OU`
attribute. When present, a worker will be restricted to only using those
host tags.

Signed-off-by: Andy Doan <andy@foundries.io>