api: Support workers with JWT auth
Add new logic to our worker authentication logic to detect a JWT. Since
JWT workers can be restricted to certain host-tags, we must also check
for that to prevent a worker from trying to "steal" work.

NOTE: There's an unauthenticate worker-create API. A rouge worker could
try to create a Worker entry here with host-tags we don't allow.
However, that will create an entry in our DB with enlisted=False. So,
this worker wouldn't have access to anything without an admin manually
enlisting it.

Signed-off-by: Andy Doan <andy@foundries.io>