runner: Use 4K reads
Our 1024 byte reads were somewhat arbitrary. They work okay on a smooth
running CI job where we are able stream the logs to the server
efficiently. However, when you start to hit timeout issues uploading,
you get a backlog of data and are stuck processing it in 1k chunks.

Our current setup allows for 4k reads. This maximizes how much we can
read and upload which should help us catch up quicker when we hit
issues.

In theory, we should also be able to do even larger reads if we set the
pipes up correctly. An 8 or 16K buffer is probably more suited to todays
internet speeds.

Signed-off-by: Andy Doan <andy@foundries.io>