github: Commit runs from a thread
This isn't super safe, but Github has limits to how long a request
can take. We have builds that now trigger 40+ CI runs. The processing
takes too long for github.

This change is the least invasive way to get what we need without having
to refactor a bunch of code. We commit the build then create a thread
to commit all the Runs and then make all the github API calls that
set the PR tests up.

Signed-off-by: Andy Doan <andy@foundries.io>