models: Optimize Build creation
By making callers do a separate commit for the trigger and reason, we
not only force 2 DB commits, our sqlalchemy session invalidates the
build object and requires a SELECT in between calls to refresh the build
id.

This isn't efficient. It's also starting to lead to an interesting
backend problem:

  sqlalchemy.orm.exc.ObjectDeletedError

This is happening when trying to commit the trigger/reason. The cause
seems to be our multi-node Galera cluster. The node we do the select on
isn't in-sync with the node we committed the Build object to.

We need to analyze our DB performance. But we also should just keep
things simpler and do an atomic operation.

Signed-off-by: Andy Doan <andy@foundries.io>