api: Optimize scale worker registration
A scale worker "registers" itself to the system by calling the
worker-update API. This API is already doing a commit on the DB, so this
call to commit is just causing extra work for something that could be
done once/atomically.

This also happens to help with an issue we've seen in production with:

 sqlalchemy.orm.exc.ObjectDeletedError

Basically we succeed with the first commit but fail on the 2nd leaving:

 * garbage in the DB
 * a worker that fails to "register"

The reason we see this is timing between Galera nodes.

Signed-off-by: Andy Doan <andy@foundries.io>