worker: Remove extra call to Build.refresh_status
Here's the history:

We originally had the worker.py code handling the DB logistics of
assigning a run to a worker. It was at this time we did the
`build.refresh_status` call in the worker logic:

 b70ed4cb944216d9163717b3d4a843cf68756367

However, this had a weird timing thing where builds would show QUEUED
for a moment when they really weren't. This was "fixed" in commit:

 4947370b4ea1a3c6f2fc4391db29171903baf0e4

However, this commit didn't remove the old call. In theory this
shouldn't be a big issue, but when the system is really busy its just
extra stuff that might cause issues. And - we are seeing some weirdness
with pop_queued and its call to build.refresh_status call.

Signed-off-by: Andy Doan <andy@foundries.io>