storage: Lazy load gcs bucket
GCP suffered a massive outage today. The thing that affected us most was
that calls to `get_bucket` were timing out. Many parts of code need a
reference to the Storage object, but they don't need the bucket. For
example, a CI run will mostly be streaming console logs to disk.

By lazy loading the bucket reference, we only pay the price of a GCS
failure in code paths that require it.

Signed-off-by: Andy Doan <andy@foundries.io>