Add logic to denote a Run is really in-progress
We've spotted the following bug happening quite frequently in
production:

 * Worker asks for work via the `GET /workers/<worker>/`
 * The `worker_get` function on the server finds a QUEUED
   Run and assigns it to the worker in our DB.
 * The `worker_get` function tries to return this to the worker
 * The worker experiences a ReadTimeout and never finds out it should be
   processing the Run
 * The Run is stuck forever.

This change introduces a `running_acked` field so we can know when the
Worker has actually started on the Run. The worker monitor logic will
then be able to detect this and fix the situation. However, that change
must be done in a separate commit so that we can deploy this safely in 2
phases.

Signed-off-by: Andy Doan <andy@foundries.io>