bump requirements
This gets us on newer versions of everything based on a similar PR to
gavelci. One dependency not touch is sqlalchemy. The v2 version of this
changes how Join operations are done which will require a small change
to our logic for run history. It will require even larger changes to our
private implemenation that imports this.

Signed-off-by: Andy Doan <andy@foundries.io>